Added transparent brick type
Added brick texture to normal bricks.
Fixed errors: Collision sound no longer plays on Awake(). Removed
delayed playback function call (i.e. audio.play(10000)).
Created transparent brick. Each brick of this type has a child
GameObject that moves with it (the cat but for now there's a placeholder
Capsule GameObject). When the brick is destroyed its child will fall due
to gravity.
Edited the scene I'm currently playtesting in ("Isaiah.unity") to have 1
of the new type of bricks. All the bricks (the original 6 + the new one)
are now subject to gravity...just because.

Destroying the child with a delay (i.e. cat falling to its death after
its brick is destroyed)--which is what we want I think--can easily
introduce memory leaks so I'm not sure yet on the best solution
especially because it doesn't seem C# supports pointers. In C++ I would
pass a pointer to the child GameObject to the GameManager instance so it
can perform the delayed destruction. But I'm not sure if that's the best
method or even necessary with C#. Any help on this is appreciated.

Gameplay suggestion: Sometimes the ball can slow down significantly so
I'm wondering if we should gradually add speed to the ball if it falls
below a certain threshold until it meets that threshold.