Created "leaping cat" transparent brick
But it does not always behave as intended. It seems to work ok when the
transparent brick is below the paddle. But seems to both go in the wrong
direction and speed when the transparent brick is above the paddle.
Sometimes, even cats below the paddle will go at too high of a speed. I
think it has to do with the "bx-by" term I derived while doing the math.
Something's wrong, so this is a rough draft.