Added simple level transition script
Mostly hardcoded atm because we don't actually have levels. The script
enables upon gameover (because it's easier to playtest losing!). It
disables objects with certain tags atm but it's a little sloppy and
later we can for example use layers corresponding to each level instead.

And tiny fix to Brick.cs:
if (other.transform.tag != "Ball" && other.transform.tag != "Weapon")
return;

"&&" instead of "||" because I think that's what we want.