Fixed some compiler warnings in the console
Some of shawn's scripts were using a private bool variable named
"enabled," but that's already a monobehavior member which enables or
disables the game object itself. So, I renamed it to "active"

Also fixed the null reference in GameManager. Now it just informs you
that you don't have the transition set up if it's not.