Added art, started making art prefabs, optimized ball script
ball script was checking for cardboard.SDK every frame, which was highly
inefficient. Now it just checks on awake and assigns it to a private
boolean.

Space and Powerup prefabs are all created. In the process of making the
others. They have a "FacePlayer" script so you never have to see the
transparent backside of them while playing.